+++ packages/opencensus-core/src/stats/recorder.ts
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
+import {TagKey, TagValue} from '../tags/types';
 import {AggregationData, AggregationType, CountData, DistributionData, LastValueData, Measurement, MeasureType, SumData} from './types';
+
+const UNKNOWN_TAG_VALUE: TagValue = null;
 
 export class Recorder {
   static addMeasurement(
@@ -38,6 +41,15 @@ export class Recorder {
       default:
         return this.addToLastValue(aggregationData, value);
     }
+  }
+
+  /** Gets the tag values from tags and columns */
+  static getTagValues(tags: Map<TagKey, TagValue>, columns: TagKey[]):
+      TagValue[] {
+    return columns.map(
+        (tagKey) =>
+            (tags.get(tagKey) ||
+             /** replace not found key values by null. */ UNKNOWN_TAG_VALUE));
   }
 
   private static addToDistribution(