+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 - Add Metrics API.
 - Add Resource API.
+- Add Tags API.
 - Add Gauges (`DoubleGauge`, `LongGauge`, `DerivedDoubleGauge`, `DerivedLongGauge`) APIs.
 - Add support for supplying instrumentation configuration via tracing option. Option argument added to instrumentation interface.
 - Add ignoreIncomingPaths and ignoreOutgoingUrls support to the http and https tracing instrumentations.
@@ -14,26 +15,82 @@ All notable changes to this project will be documented in this file.
 
 - Modify `Logger` interface: `level` made optional, `silly` removed.
 - The ```new Stats()``` has been deprecated on Stats class. The global singleton ```globalStats``` object should be used instead. Also, ```registerView()``` is separated out from ```createView()```.
+- Use ```TagKey```, ```TagValue``` and ```TagMap``` to create the tag keys, tag values.
 
 ##### Old code
 ```js
 const { Stats } = require("@opencensus/core");
 const stats = new Stats();
 
+// Counts/groups the lengths of lines read in.
+const mLineLengths = stats.createMeasureInt64(
+  "demo/line_lengths",
+  MeasureUnit.BYTE,
+  "The distribution of line lengths"
+);
+
+// Create tag keys
+const tagKeys = ["method", "status"];
+
 // Create and register the view
-stats.createView(...);
+stats.createView(
+  "demo/lines_in",
+  mLineLengths,
+  AggregationType.COUNT,
+  tagKeys,
+  "The number of lines from standard input"
+);
+
+// Records measurements
+stats.record({
+  measure: mLineLengths,
+  tags,
+  value: 2
+});
+
 ```
 
 ##### New code
 ```js
-// Get the global singleton stats object
+// Gets the global stats instance
 const { globalStats } = require("@opencensus/core");
 
-// Create the view
-const view = globalStats.createView(...);
+// Counts/groups the lengths of lines read in.
+const mLineLengths = globalStats.createMeasureInt64(
+  "demo/line_lengths",
+  MeasureUnit.BYTE,
+  "The distribution of line lengths"
+);
 
-// register the view
+// Creates the method and status key
+const methodKey = {name: "method"};
+const statusKey = {name: "status"};
+
+// Creates the view
+const view = globalStats.createView(
+  "demo/lines_in",
+  mLineLengths,
+  AggregationType.COUNT,
+  [methodKey, statusKey],
+  "The number of lines from standard input"
+);
+
+// Registers the view
 globalStats.registerView(view);
+
+// Creates tags map -> key/value pair
+const tagMap = new TagMap();
+tagMap.set(methodKey, {value: 'REPL'});
+tagMap.set(statusKey, {value: 'OK'});
+
+// Creates measurements (measure + value)
+const measurements = [{
+  measure: mLineLengths,
+  value: 2
+}];
+
+// Records measurement with tagMap
+globalStats.record(measurements, tagMap);
 ```
 
 ## 0.0.8 - 2018-12-14